Refactor service to use new models
