+++ app/services/idm/fetch-user-roles-and-groups.service.js
@@ -5,7 +5,7 @@
  * @module FetchUserRolesAndGroupsService
  */
 
-const UserModel = require('../../models/idm/user.model.js')
+const UserModel = require('../../models/user.model.js')
 
 /**
  * A user can have roles assigned to them in two ways:
@@ -77,13 +77,13 @@ function _extractRolesFromGroups (groups) {
 }
 
 function _combineAndDedupeRoles (rolesArrayToDedupe) {
-  // Our usual method of deduping arrays (putting the array into a new Set and then spreading it back into an array)
+  // Our usual method of de-duping arrays (putting the array into a new Set and then spreading it back into an array)
   // doesn't work here as the Role objects are not considered to be equal when doing this. We therefore use reduce to
   // dedupe by going through each Role object in the original array and only adding it to the accumulated results array
   // if a Role object with the same id isn't already in it
   const dedupedArray = rolesArrayToDedupe.reduce((acc, current) => {
     const roleIsAlreadyInAcc = acc.find((item) => {
-      return item.roleId === current.roleId
+      return item.id === current.id
     })
 
     if (!roleIsAlreadyInAcc) {
